chat/requirements.txt
langchain
openai
urllib3==1.26.13
pinecone-client